+++ bin/metrics-beanstalkd-tubes.rb
@@ -73,6 +73,15 @@ class BeanstalkdMetrics < Sensu::Plugin::Metric::CLI::Graphite
 
   INGORED_KEYS = ['name'].freeze
 
+  def tube_age(tube)
+    job = tube.peek(:ready)
+    if !job.nil?
+      job.stats.age
+    else
+      0
+    end
+  end
+
   def output_stats(tube)
     stats = tube.stats
     stats.keys.sort.each do |key|
@@ -80,6 +89,7 @@ class BeanstalkdMetrics < Sensu::Plugin::Metric::CLI::Graphite
       next if INGORED_KEYS.include?(key)
       output "#{config[:scheme]}.#{tube.name}.#{key}", stats[key]
     end
+    output "#{config[:scheme]}.#{tube.name}.age", tube_age(tube) if matches_filter?(:stats, 'age')
   end
 
   def tubes